Move scheme classes into JMeter's package namespace.
